Add firebase.js and env.local
