+++ Frontend/User_Authentication/package.json
@@ -11,7 +11,10 @@
   },
   "dependencies": {
     "axios": "^1.7.9",
+    "bootstrap": "^5.3.3",
+    "firebase": "^11.2.0",
     "react": "^18.3.1",
+    "react-bootstrap": "^2.10.9",
     "react-dom": "^18.3.1"
   },
   "devDependencies": {
