+++ Frontend/User_Authentication/src/Components/App.jsx
@@ -1,0 +1,13 @@
+import React from "react";
+import AuthForm from "./AuthForm";
+import "./App.css";
+
+const App = () => {
+  return (
+    <div className="OuterContainer">
+      <AuthForm />
+    </div>
+  );
+};
+
+export default App;
